+++ src/main/java/com/t3t/bookstoreapi/book/repository/BookRepositoryCustom.java
@@ -3,7 +3,7 @@ package com.t3t.bookstoreapi.book.repository;
 import com.t3t.bookstoreapi.book.model.dto.CategoryDto;
 import com.t3t.bookstoreapi.book.model.dto.ParticipantRoleRegistrationDto;
 import com.t3t.bookstoreapi.book.model.dto.ParticipantRoleRegistrationDtoByBookId;
-import com.t3t.bookstoreapi.book.model.dto.TagDto;
+import com.t3t.bookstoreapi.tag.model.dto.TagDto;
 import com.t3t.bookstoreapi.book.model.response.BookDetailResponse;
 import com.t3t.bookstoreapi.recommendation.model.response.BookInfoBriefResponse;
